+++ scripts/bundle.msix.ts
@@ -64,6 +64,10 @@ if (pre || target === "./") {
 // dlls
 fs.copyFileSync(`target/${target}/release/sluhk.dll`, `${buildFolder}/sluhk.dll`);
 
+// integrity files
+fs.copyFileSync(`target/${target}/release/SHA256SUMS`, `${buildFolder}/SHA256SUMS`);
+fs.copyFileSync(`target/${target}/release/SHA256SUMS.sig`, `${buildFolder}/SHA256SUMS.sig`);
+
 // Add resources
 fs.cpSync("src/static", `${buildFolder}/static`, { recursive: true });