+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [2.5.1-dev]
+## [2.5.1]
 
 ### enhancements